requirements.txt
langchain>=0.0.148
langchain_ollama
langchain_community
googlesearch-python